Fixes associated type not found in libraries.
This error occured because `TypeCheckContext` resolve was calling
`self.namespace.root.resolve_call_path_and_root_type_id` instead of `self.namespace.resolve_call_path_and_root_type_id`.

For this to work the `Root` resolve methods were moved to `Module`.

Closes #5525
