+++ test/src/e2e_vm_tests/test_programs/should_pass/language/associated_type_container_in_library/Forc.lock
@@ -1,0 +1,13 @@
+[[package]]
+name = "associated_type_container_in_library"
+source = "member"
+dependencies = ["std"]
+
+[[package]]
+name = "core"
+source = "path+from-root-2E0F43F5F5F56DC3"
+
+[[package]]
+name = "std"
+source = "path+from-root-2E0F43F5F5F56DC3"
+dependencies = ["core"]
