+++ test/src/e2e_vm_tests/test_programs/should_pass/language/associated_type_container_in_library/test.toml
@@ -1,0 +1,3 @@
+category = "run"
+expected_result = { action = "return", value = 1 }
+expected_warnings = 1
